+++ tests/TestCase/ExceptionistTest.php
@@ -182,9 +182,9 @@ class ExceptionistTest extends TestCase
             [0, 'Value `0` is not equal to `true`'],
         ] as $exception) {
             [$value, $expectedMessage] = $exception;
-            $this->assertException(Exception::class, function () use ($value) {
+            $this->assertException(function () use ($value) {
                 Exceptionist::isTrue($value);
-            }, $expectedMessage);
+            }, Exception::class, $expectedMessage);
         }
     }
 
@@ -196,17 +196,17 @@ class ExceptionistTest extends TestCase
     {
         $message = 'it\'s not `true`';
 
-        $this->assertException(Exception::class, function () use ($message) {
+        $this->assertException(function () use ($message) {
             Exceptionist::isTrue(false, $message);
-        }, $message);
+        }, Exception::class, $message);
 
-        $this->assertException(ErrorException::class, function () use ($message) {
+        $this->assertException(function () use ($message) {
             Exceptionist::isTrue(false, new ErrorException($message));
-        }, $message);
+        }, ErrorException::class, $message);
 
-        $this->assertException(ErrorException::class, function () use ($message) {
+        $this->assertException(function () use ($message) {
             Exceptionist::isTrue(false, $message, ErrorException::class);
-        }, $message);
+        }, ErrorException::class, $message);
     }
 
     /**
@@ -215,8 +215,8 @@ class ExceptionistTest extends TestCase
      */
     public function testIsTrueFailureWithInvalidExceptionClass()
     {
-        $this->assertException(Notice::class, function () {
+        $this->assertException(function () {
             Exceptionist::isTrue(false, '', new \stdClass());
-        }, '`$exception` parameter must be an instance of `Throwable` or a string');
+        }, Notice::class, '`$exception` parameter must be an instance of `Throwable` or a string');
     }
 }