+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -61,15 +61,17 @@ class TestTraitTest extends TestCase
         }
 
         //Missing argument
-        $this->assertException(BadMethodCallException::class, function () {
-            $this->assertIsJson();
-        }, 'Method ' . get_parent_class($this) . '::assertIsJson() expects at least 1 argument, maximum 2, 0 passed');
+        $this->assertException(
+            [$this, 'assertIsJson'],
+            BadMethodCallException::class,
+            'Method ' . get_parent_class($this) . '::assertIsJson() expects at least 1 argument, maximum 2, 0 passed'
+        );
 
         //Calling a no existing method or a no existing "assertIs" method
         foreach (['assertIsNoExistingType', 'noExistingMethod'] as $method) {
-            $this->assertException(BadMethodCallException::class, function () use ($method) {
+            $this->assertException(function () use ($method) {
                 $this->$method('string');
-            }, 'Method ' . get_parent_class($this) . '::' . $method . '() does not exist');
+            }, BadMethodCallException::class, 'Method ' . get_parent_class($this) . '::' . $method . '() does not exist');
         }
     }
 
@@ -98,19 +100,19 @@ class TestTraitTest extends TestCase
      */
     public function testAssertException()
     {
-        $this->assertException(Exception::class, function () {
+        $this->assertException(function () {
             throw new Exception();
         });
-        $this->assertException(Exception::class, function () {
+        $this->assertException(function () {
             throw new Exception('right exception message');
         });
-        $this->assertException(Exception::class, function () {
+        $this->assertException(function () {
             throw new Exception('right exception message');
-        }, 'right exception message');
+        }, Exception::class, 'right exception message');
 
         //No exception throw
         try {
-            $this->assertException(Exception::class, 'time');
+            $this->assertException('time');
         } catch (AssertionFailedError $e) {
             $this->assertStringStartsWith('Expected exception `Exception`, but no exception throw', $e->getMessage());
         } finally {
@@ -123,9 +125,9 @@ class TestTraitTest extends TestCase
         //No existing exception or invalid exception class
         foreach (['noExistingException', stdClass::class] as $class) {
             try {
-                $this->assertException($class, function () {
+                $this->assertException(function () {
                     throw new Exception();
-                });
+                }, $class);
             } catch (AssertionFailedError $e) {
                 $this->assertStringStartsWith('Class `' . $class . '` does not exist or is not an exception', $e->getMessage());
             } finally {
@@ -138,9 +140,9 @@ class TestTraitTest extends TestCase
 
         //Unexpected exception type
         try {
-            $this->assertException(Deprecated::class, function () {
+            $this->assertException(function () {
                 throw new Exception();
-            });
+            }, Deprecated::class);
         } catch (AssertionFailedError $e) {
             $this->assertStringStartsWith('Expected exception `' . Deprecated::class . '`, unexpected type `Exception`', $e->getMessage());
         } finally {
@@ -152,9 +154,9 @@ class TestTraitTest extends TestCase
 
         //Wrong exception message
         try {
-            $this->assertException(Exception::class, function () {
+            $this->assertException(function () {
                 throw new Exception('Wrong');
-            }, 'Right');
+            }, Exception::class, 'Right');
         } catch (AssertionFailedError $e) {
             $this->assertStringStartsWith('Expected message exception `Right`, unexpected message `Wrong`', $e->getMessage());
         } finally {
@@ -166,9 +168,9 @@ class TestTraitTest extends TestCase
 
         //Expected exception message, but no message
         try {
-            $this->assertException(Exception::class, function () {
+            $this->assertException(function () {
                 throw new Exception();
-            }, 'Right');
+            }, Exception::class, 'Right');
         } catch (AssertionFailedError $e) {
             $this->assertStringStartsWith('Expected message exception `Right`, but no message for the exception', $e->getMessage());
         } finally {
@@ -242,9 +244,9 @@ class TestTraitTest extends TestCase
             [null],
             [''],
         ] as $array) {
-            $this->assertException(AssertionFailedError::class, function () use ($array) {
+            $this->assertException(function () use ($array) {
                 $this->assertIsArrayNotEmpty($array);
-            });
+            }, AssertionFailedError::class);
         }
     }